requirements.txt
requests
paho-mqtt
python-dotenv
selenium
webdriver-manager
flask
flask-cors
Flask-SocketIO
gevent-websocket
pycurl
